requirements.txt
Django==1.9.2
django-appconf==1.0.1
django-axes==1.6.0
django-bower==5.1.0
django-compat==1.0.8
django-compressor==2.0
django-crispy-forms==1.6.0
django-enumfield==1.2.1
django-extensions==1.6.1
django-hijack==2.0.3
django-jet==0.1.4
django-localflavor==1.2
django-model-utils==2.4
django-sass-processor==0.3.4
libsass==0.10.1
Pint==0.7.1
pytz==2015.7
rcssmin==1.0.6
rjsmin==1.0.12
rules==1.1.1
six==1.10.0
